finaly i fixed the refactoring

+++ src/basic_config.h
@@ -2,15 +2,15 @@
 
 struct BasicConfig
 {
-    float WINDOW_WIDTH;
-    float WINDOW_HEIGHT;
+    int WINDOW_WIDTH;
+    int WINDOW_HEIGHT;
     int FPS;
     float SIZE_MULTIPLIER;
 };
 
 static const BasicConfig basic_config = {
-    .WINDOW_WIDTH = 800.0f,
-    .WINDOW_HEIGHT = 600.0f,
+    .WINDOW_WIDTH = 800,
+    .WINDOW_HEIGHT = 600,
     .FPS = 60,
     .SIZE_MULTIPLIER = 1,
 };
